feat: move CollectionCalendarView + CollectionDayView into collection-plugin/vue (Phase 2 step 2c)
Two more pure leaves (useI18n-only) onto the SFC pipeline. The month-grid
calendar view and the single-day lane/timeline view move into src/vue/components/,
with their host couplings rebound to package-internal relative paths:

- enumColors (resolveEnumColor/EnumColorClasses) → ../../core/enumColors
- collectionTypes (CollectionItem/CollectionSchema) → ../../core/schema
- calendar engine (bucketRecords/buildMonthGrid/daySlice/assignLanes/…) and
  itemLabel helpers: the `@mulmoclaude/collection-plugin` index imports become
  relative ../../core/{calendarGrid,itemLabel} — a package can't import its own
  published entry mid-build. useI18n stays (vue-i18n peer).

CollectionView imports both from @mulmoclaude/collection-plugin/vue (folded into
the existing CollectionRecordModal import).

Verified: package build (style.css 31.9kB, vue.js 38.2kB) + package vue-tsc +
host typecheck green.
