+++ packages/plugins/collection-plugin/src/vue/components/CollectionDayView.vue
@@ -115,10 +115,10 @@
 <script setup lang="ts">
 import { computed, nextTick, onMounted, ref } from "vue";
 import { useI18n } from "vue-i18n";
-import { bucketRecords, daySlice, assignLanes, ymdKey, MINUTES_PER_DAY, type Ymd, type DaySlice } from "@mulmoclaude/collection-plugin";
-import { resolveEnumColor, type EnumColorClasses } from "../utils/collections/enumColors";
-import { labelFieldFor, itemIdOf, itemLabelOf } from "@mulmoclaude/collection-plugin";
-import type { CollectionItem, CollectionSchema } from "./collectionTypes";
+import { bucketRecords, daySlice, assignLanes, ymdKey, MINUTES_PER_DAY, type Ymd, type DaySlice } from "../../core/calendarGrid";
+import { resolveEnumColor, type EnumColorClasses } from "../../core/enumColors";
+import { labelFieldFor, itemIdOf, itemLabelOf } from "../../core/itemLabel";
+import type { CollectionItem, CollectionSchema } from "../../core/schema";
 
 const props = defineProps<{
   schema: CollectionSchema;
