+++ src/components/CollectionView.vue
@@ -727,10 +727,8 @@ import { BUILTIN_ROLE_IDS } from "../config/roles";
 import ConfirmModal from "./ConfirmModal.vue";
 import PinToggle from "./PinToggle.vue";
 import CollectionRecordPanel from "./CollectionRecordPanel.vue";
-import { CollectionRecordModal } from "@mulmoclaude/collection-plugin/vue";
+import { CollectionRecordModal, CollectionCalendarView, CollectionDayView } from "@mulmoclaude/collection-plugin/vue";
 import CollectionViewConfigModal from "./CollectionViewConfigModal.vue";
-import CollectionCalendarView from "./CollectionCalendarView.vue";
-import CollectionDayView from "./CollectionDayView.vue";
 import CollectionKanbanView from "./CollectionKanbanView.vue";
 import CollectionCustomView from "./CollectionCustomView.vue";
 import {
